requirements.txt
paho-mqtt
boto3
mysql-connector-python
pandas
numpy
scikit-learn
tensorflow
flask

